requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.1.1
Deprecated==1.2.13
Django==4.1
django-filter==22.1
django-jazzmin==2.5.0
djangorestframework==3.13.1
heroku==0.1.4
heroku3==5.1.4
idna==3.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
